fix: Remove some configurations on Admin and Update template email

+++ microservice-auth/src/main/java/com/microservice/auth/repositories/UserRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<AuthUser, Integer> {
+public interface UserRepository extends JpaRepository<AuthUser, Long> {
     Optional<AuthUser> findByEmail(String email);
 }
